Introduce new grid method: Grid::GetNumDimensions() (#2752)
* finish surgeries on grids

* move member data from private to protected

* clang-format pre-push hook

* same safe work

* massive API change

* clang-format pre-push hook

* minor change

* clang-format pre-push hook

* more changes

* clang-format pre-push hook

* work on StructuredGrid

* clang-format pre-push hook

* fix typo

* minor

* clang-format pre-push hook

* minor

* clang-format pre-push hook

* minor

* clang-format pre-push hook

* minor

* clang-format pre-push hook

* minor

* clang-format pre-push hook

* minor

* minor

* clang-format pre-push hook

* minor

* clang-format pre-push hook

* minor

* clang-format pre-push hook

* minor

* clang-format pre-push hook

* minor

* clang-format pre-push hook

* minor

* minor

* minor

* clang-format pre-push hook

* minor

* clang-format pre-push hook

* fix text programs

* clang-format pre-push hook

* minor

* clang-format pre-push hook

* remove reference to GrownGrid

* clang-format pre-push hook

* minor

* remove a few comments

* format

* touch Grid.h and ConstantGrid.h

* minor

* minor

* minor

* minor

* minor

* minor

* minor

* minor

* minor

* CurvilinearGrid.cpp

* minor

* minor

* Grid

* Grid

* StretchedGrid

* StructuredGrid

* UnstructuredGrid.cpp

* UnstructuredGrid2D

* lib/vdc/UnstructuredGrid2D.cpp

* lib/vdc/UnstructuredGrid2D.cpp

* UnstructuredGrid3D.cpp

* UnstructuredGridLayered

* test apps

* clang-format

* fix conflict

* minor

* minor

* minor

* minor

* minor

* minor

* clang-format

* minor

* performance improv

* clang-format

* revert bad cmake

* improve CurvilinearGrid

* minor

* minor

* minor

* Introduce PAxisAnnotationWidget - Resolve Issue2724 (#2753)

* inc - build broken

* debugging PCS issue

* revert AnnotationRenderer and add new files

* remove unused code

* address stas' suggestions

* reformat

* fix 2758 (#2762)

* minor

* clang-format

+++ include/vapor/utils.h
@@ -65,7 +65,7 @@ private:
 //! \retval offset The offset from the first element of the array to the
 //! address specified by \p coords
 //
-COMMON_API size_t LinearizeCoords(const size_t *coords, const size_t *dims, int n);
+COMMON_API size_t LinearizeCoords(const size_t *coords, const size_t *dims, size_t n);
 COMMON_API size_t LinearizeCoords(const std::vector<size_t> &coords, const std::vector<size_t> &dims);
 
 //! Linearize multi-dimensional coordinates
@@ -82,7 +82,7 @@ COMMON_API size_t LinearizeCoords(const std::vector<size_t> &coords, const std::
 //!
 //! \retval offset The offset from the first element of the array to the
 //! address specified by \p coords
-COMMON_API size_t LinearizeCoords(const size_t *coords, const size_t *min, const size_t *max, int n);
+COMMON_API size_t LinearizeCoords(const size_t *coords, const size_t *min, const size_t *max, size_t n);
 COMMON_API size_t LinearizeCoords(const std::vector<size_t> &coords, const std::vector<size_t> &min, const std::vector<size_t> &max);
 
 //! Increment a coordinate vector by one
@@ -106,12 +106,12 @@ COMMON_API size_t VProduct(const std::vector<size_t> &a);
 
 //! Vectorize a coordinate offset. Inverse of VectorizeLinearize
 //!
-COMMON_API void VectorizeCoords(size_t offset, const size_t *min, const size_t *max, size_t *coords, int n);
+COMMON_API void VectorizeCoords(size_t offset, const size_t *min, const size_t *max, size_t *coords, size_t n);
 COMMON_API std::vector<size_t> VectorizeCoords(size_t offset, const std::vector<size_t> &min, const std::vector<size_t> &max);
 
 //! Vectorize a coordinate offset. Inverse of VectorizeLinearize
 //!
-COMMON_API void VectorizeCoords(size_t offset, const size_t *dims, size_t *coords, int n);
+COMMON_API void VectorizeCoords(size_t offset, const size_t *dims, size_t *coords, size_t n);
 COMMON_API std::vector<size_t> VectorizeCoords(size_t offset, const std::vector<size_t> &dims);
 
 //
